update budget for user

+++ src/budgets/repository/budgets.repository.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { Budget } from "src/generated/prisma/client";
 import { PrismaService } from "src/prisma/prisma.service";
 
@@ -8,30 +8,47 @@ import { UpdateBudgetRequest } from "../dto/update-budget.dto";
 @Injectable()
 export class BudgetsRepository {
   constructor(private readonly prisma: PrismaService) {}
-
-  async getAll(): Promise<Budget[]> {
+  private connectUser(userId: string) {
+    return { connect: { id: userId } };
+  }
+  async getAll(userId: string): Promise<Budget[]> {
     return this.prisma.budget.findMany({
+      where: { userId },
       include: { categories: true },
-      orderBy: { label: "asc" }, // Trier par label par ordre alphabétique
+      orderBy: { label: "asc" },
     });
   }
 
-  async create(budget: CreateBudgetRequest): Promise<Budget> {
-    return this.prisma.budget.create({ data: budget });
+  async create(userId: string, budget: CreateBudgetRequest): Promise<Budget> {
+    return this.prisma.budget.create({
+      data: {
+        ...budget,
+        user: this.connectUser(userId),
+      },
+    });
   }
 
-  getById(id: string): Promise<Budget | null> {
+  getById(userId: string, id: string): Promise<Budget | null> {
     return this.prisma.budget.findUnique({
-      where: { id },
+      where: { id, userId },
       include: { categories: true },
     });
   }
 
-  updateOne(id: string, budget: UpdateBudgetRequest): Promise<Budget> {
+  async updateOne(userId: string, id: string, budget: UpdateBudgetRequest): Promise<Budget> {
+    const existingBudget = await this.getById(userId, id);
+    if (!existingBudget) {
+      throw new NotFoundException(`Budget with ID ${id} not found`);
+    }
     return this.prisma.budget.update({ where: { id }, data: budget });
   }
 
-  remove(id: string): Promise<Budget> {
+  async remove(userId: string, id: string): Promise<Budget> {
+    const existingBudget = await this.getById(userId, id);
+
+    if (!existingBudget) {
+      throw new NotFoundException(`Budget with ID ${id} not found`);
+    }
     return this.prisma.budget.delete({ where: { id } });
   }
 }
